Sink type refactoring
This commit is motivated by the fact that the `linkMerge` and `pickValue`
fields were repeated in the `Sink` and `WorkflowOutputParameter` types.

To remove this repetition, these two fields have been isolated in a new
type called `Sink`. The previous `Sink` type, which is now named `InputSink`
extends `Sink` with the `source` field, while a new `OutputSink` type extends
`Sink` with the `outputSource` field.

The `WorkflowInputParameter` type extends the `InputSink`, while the
`WorkflowOutputParameter` type inherits from `OutputSink`. Note that
this separation will be useful also for future extensions. For example,
the proposed `loop` feature will build upon the `OutputSink` type to map
the step outputs into the inputs of the next iteration.